requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.1
amqp==5.3.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==2.1.0
asgiref==3.8.1
asto==0.0.34
astr==0.0.11
asyncpg==0.30.0
attrs==24.2.0
atup==0.0.12
atyp==0.0.62
BaseML==0.1.5
bflag==0.0.9
billiard==4.2.1
bitarray==3.0.0
bitstring==4.3.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chck==0.0.75
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cron-descriptor==1.4.5
cycler==0.12.1
Django==5.1.3
django-celery-beat==2.7.0
django-timezone-field==7.0
dnspython==2.7.0
dref==0.0.1
eventlet==0.37.0
face_sdk_3divi @ file:///home/slonyara/3DiVi_FaceSDK/3_24_2/python_api
fastapi==0.115.5
fonttools==4.55.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idfunc==0.0.8
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
ispec==0.0.18
itry==0.0.18
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.4.2
lazy_loader==0.4
ltyp==0.0.29
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.0
matr==0.0.10
minio==7.2.14
multidict==6.1.0
multipledispatch==1.0.0
nchr==0.0.11
networkx==3.4.2
nist==0.0.3
nlit==0.0.97
numpy==1.26.0
opencv-python==4.8.0.76
packaging==24.2
pandas==2.2.3
pdec==0.0.4
pfmt==0.0.4
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
pstr==0.0.49
psycopg2-binary==2.9.10
pttn==0.0.8
ptyp==0.0.19
putl==0.0.15
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
quac==0.0.45
regsyn==0.0.3
requests==2.32.3
ropts==0.0.4
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
sigr==0.0.12
sint==0.0.10
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlitedict==2.1.0
sqlparse==0.5.2
starlette==0.41.2
threadpoolctl==3.5.0
tifffile==2025.1.10
typing_extensions==4.12.2
typs==0.0.6
tzdata==2024.2
urllib3==2.2.3
uscr==0.0.2
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1
yellowbrick==1.5
